tg FTCGIT/requirements.txt
python-telegram-bot>=21.0
google-genai>=0.1.0
python-dotenv==1.0.0
